automation_hub/apps/api/requirements.txt
# ============ Core ============
fastapi==0.115.0
uvicorn[standard]==0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ============ Database ============
supabase>=2.3.0

# ============ Scheduler ============
apscheduler>=3.10.0

# ============ Utils ============
python-dotenv>=1.0.0
python-multipart>=0.0.6

# ============ Testing ============
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
